mandrill-webhook/requirements.txt
fastapi            == 0.94.1
pymongo            == 4.3.3
uvicorn            == 0.21.1
PyYAML             == 6.0
python-socketio    == 5.8.0